# docs/API_SPECIFICATION.yaml
openapi: 3.0.3
info:
  title: Brix Nutritional App API
  version: 2.0.0
  description: |
    REST API для Brix Nutritional App (Mobile + Admin Web Panel).
    
    ## Компоненты системы
    
    ### 1. Mobile App (Flutter)
    Мобильное приложение для пользователей (iOS/Android).
    Использует все endpoints из категорий: Auth, Users, Home, Meal Plans, Recipes, Diary, AI Chat, Lab Tests, Courses, Blog, Notifications, Subscriptions.
    
    ### 2. Admin Web Panel (Next.js)
    Веб-панель для администраторов и контент-менеджеров.
    Использует admin-специфичные endpoints и endpoints для управления контентом (Courses, Lessons, Recipes, Meal Plans, Users, Blog, Subscriptions, Analytics).
    
    ### 3. Backend API (Strapi)
    Сервер на Node.js + Strapi CMS с кастомными контроллерами.
    
    ---
    
    ## Авторизация
    
    ### Для Mobile App:
    Большинство endpoints требуют JWT токен в заголовке:
    ```
    Authorization: Bearer <token>
    ```
    
    ### Для Admin Web Panel:
    Admin endpoints требуют Strapi admin JWT токен:
    ```
    Authorization: Bearer <admin-token>
    ```
    Токен получается через Strapi `/admin/login` endpoint.
    
    ---
    
    ## Rate Limiting
    - Auth endpoints: 5 запросов / 15 минут
    - Mobile App endpoints: 100 запросов / минуту
    - Admin endpoints: 1000 запросов / минуту
    
  contact:
    name: API Support
    email: support@brix-nutrition.com
servers:
  - url: http://localhost:1337/api
    description: Development
...
paths:
...
  /lab-tests/upload:
    post:
      tags: [Lab Tests]
      summary: Загрузить результаты анализов
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [testType, testDate, results]
              properties:
                testType:
                  type: string
                  example: "blood_general"
                testDate:
                  type: string
                  format: date
                results:
                  type: array
                  items:
                    type: object
                    required: [parameterId, value, unit]
                    properties:
                      parameterId:
                        type: string
                        example: "HGB"
                      value:
                        type: number
                        example: 128
                      unit:
                        type: string
                        example: "г/л"
      responses:
        '201':
          description: Анализ загружен
          content:
            application/json:
              schema:
                type: object
                properties:
                  testId:
                    type: string
                    format: uuid
  /lab-tests/my:
    get:
      tags: [Lab Tests]
      summary: Получить мои анализы
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Список анализов
          content:
            application/json:
              schema:
                type: object
                properties:
                  tests:
                    type: array
                    items:
                      $ref: '#/components/schemas/LabTest'
  /lab-tests/{id}:
    get:
      tags: [Lab Tests]
      summary: Получить детали анализа
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Детали
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LabTest'
...
  /lab-tests/parameters:
    get:
      tags: [Lab Tests]
      summary: Получить справочник показателей
      security:
        - bearerAuth: []
      parameters:
        - name: category
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Справочник
          content:
            application/json:
              schema:
                type: object
                properties:
                  parameters:
                    type: array
                    items:
                      $ref: '#/components/schemas/LabParameter'

  # ============= COURSES =============
...
  /blog/articles:
    get:
      tags: [Blog]
      summary: Получить список статей
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: category
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Статьи
          content:
            application/json:
              schema:
                type: object
                properties:
                  articles:
                    type: array
                    items:
                      $ref: '#/components/schemas/Article'
                  total:
                    type: integer
  /blog/articles/{id}:
    get:
      tags: [Blog]
      summary: Получить статью
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Статья
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Article'

  # ============= NOTIFICATIONS =============
  /notifications:
    get:
      tags: [Notifications]
      summary: Получить уведомления
      security:
        - bearerAuth: []
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Уведомления
          content:
            application/json:
              schema:
                type: object
                properties:
                  notifications:
                    type: array
                    items:
                      $ref: '#/components/schemas/Notification'
  /notifications/{id}/read:
    patch:
      tags: [Notifications]
      summary: Отметить уведомление прочитанным
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Отмечено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /notifications/{id}:
    delete:
      tags: [Notifications]
      summary: Удалить уведомление
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Удалено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'

  # ============= SUBSCRIPTIONS =============
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
...
    LabTest:
      type: object
      properties:
        id:
          type: string
          format: uuid
        testType:
          type: string
        testDate:
          type: string
          format: date
        results:
          type: array
          items:
            type: object
            properties:
              parameterId:
                type: string
              value:
                type: number
              unit:
                type: string
    LabParameter:
      type: object
      properties:
        id:
          type: string
          example: "HGB"
        name:
          type: string
          example: "Гемоглобин"
        category:
          type: string
          example: "Клинический анализ крови"
        units:
          type: array
          items:
            type: string
        referenceRanges:
          type: array
          items:
            type: object
            properties:
              gender:
                type: string
              ageMin:
                type: integer
              ageMax:
                type: integer
              min:
                type: number
              max:
                type: number
              unit:
                type: string
...
    Article:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        slug:
          type: string
        content:
          type: string
          description: Markdown
        preview:
          type: string
        imageUrl:
          type: string
        author:
          type: string
        category:
          type: string
        publishedAt:
          type: string
          format: date-time
    Notification:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        message:
          type: string
        type:
          type: string
          enum: [info, reminder, alert]
        isRead:
          type: boolean
        action:
          type: object
          properties:
            type:
              type: string
            target:
              type: string
        createdAt:
          type: string
          format: date-time
...